+++ banyand/liaison/grpc/measure.go
@@ -215,7 +215,20 @@ func (ms *measureService) Query(ctx context.Context, req *measurev1.QueryRequest
 			span.Stop()
 		}()
 	}
-	feat, err := ms.broadcaster.Publish(ctx, data.TopicMeasureQuery, bus.NewMessage(bus.MessageID(now.UnixNano()), req))
+	nodeSelectors := make(map[string]string)
+	for _, g := range req.Groups {
+		if req.NodeSelector != "" {
+			nodeSelectors[g] = req.NodeSelector
+			continue
+		}
+		if ns, exist := ms.groupRepo.getNodeSelector(g); exist {
+			nodeSelectors[g] = ns
+			continue
+		}
+		nodeSelectors[g] = ""
+	}
+
+	feat, err := ms.broadcaster.Publish(ctx, data.TopicMeasureQuery, bus.NewMessageWithNodeSelectors(bus.MessageID(now.UnixNano()), nodeSelectors, req.TimeRange, req))
 	if err != nil {
 		return nil, err
 	}